pyproject.toml
[project]
name = "ai-travel-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "python-dotenv>=1.2.2",
    "typing-extensions>=4.15.0",
]
